nios2: (nommu) cache_clear() shouldn't write back
This patch is for the test-nios2 (nommu) branch.

The static inline function cache_invalidate_data() used internally in
arch/mm/memory.c is misnamed because it writes back the dirty cache
lines first.  Rename it to cache_push_invalidate_data() and change (most
of) the calls to it to use the new name.

Add a new version of cache_invalidate_data() which merely invalidates
the data cache lines without writing them back first.  Call this
function from the externally linked cache_clear() function.  This is
in keeping with the cache_clear() semantics.

Signed-off-by: Ian Abbott <abbotti@mev.co.uk>
Signed-off-by: Tobias Klauser <klto@zhaw.ch>